Testing ray functionality
- Added pseudo_db
	This provides functionality for basic searches and just
	generally indexing for tests of ray containers.
- Populated pseudo_db with syllabary information
- Basic word parsing for language model

Appears that the same ray layout can be used for higher order data,
although this may not be desireable for actual implementation.

It is also a decent idea to gather some source information from the
target language in order to generate a colleciton of non-word characters
or punctuation. Seperating words and thoughts may be more difficult
without seperatable characters for tokenization.