Added temp implementation of following_plus
Currently it is written to not replace following, but it probably
should to simplify things.
The major difference with this structure is it has a conditional
relationship using the degree field to either reference itself or
following table.
This should be moved to work as as an internal system and should replace
following where degrees of 1 reference instances.